extracts interface for report store
